feature and bug fix
Some tables were not shown properly in the forum when placed in the forum macro.
All local macros can now be inserted in to any forum macro.
colors were added to main edit box.
Options for to allow empty rows in tables and lists
table width added to all tables
This is (v) 2.4