feature
remove supplemental signature feature.  Instead a sentence is added to printer macros files as a suffix to the macro.

updating rtf help files in sources now updates the corresponding release and binary files if they are older.  Updating source rtf files from the release folder are done afterwards (as was previously done).